plugins/kitaku-route/requirements.txt
staticmap
fpdf2
